rename: SideBar.tsx→MenuBar.tsx
