+++ components/index.ts
@@ -1,2 +1,2 @@
-export { SideBar } from "./SideBar";
+export { SideBar } from "./MenuBar";
 export { Icon } from "./Icon";
